[moment-timezone] Fixed moment-timezone/moment-timezone import, rearranged exports
